requirements.txt
torch==1.13.1
torchvision==0.14.1
numpy==1.21.6
Pillow==9.4.0
tqdm==4.65.0
opencv-python==4.7.0
--extra-index-url https://pypi.ngc.nvidia.com
git+https://github.com/usuyama/pytorch-unet.git